feat: ナビゲーションを3項目に整理しカテゴリ管理をSettingに内包 (#30)
- ナビをDashboard→Expense→Settingの3項目に変更
- フッターのFAB特別処理を削除し均等表示に統一
- カテゴリ管理をSettingsPageにセクションとして組み込み
- CategoriesPage・/categoryルートを削除
